Storing functions was moved in enStoring
